Leader spinner added 2.Detection result added
