+++ package.json
@@ -16,6 +16,8 @@
     "react-bootstrap": "^1.0.0-beta.17",
     "react-dom": "^16.13.0",
     "react-easy-crop": "^2.1.1",
+    "react-loader-spinner": "^3.1.5",
+    "react-promise-tracker": "^2.1.0",
     "react-scripts": "3.4.0"
   },
   "scripts": {
