--- src/components/Detector/Detector.css
@@ -1,34 +1,0 @@
-#detector {
-    position: relative;
-}
-.Result {
-    margin-bottom: 20px;
-    text-align: center;
-    font-weight: 700;
-    font-family: Poppins, Arial, sans-serif;
-    -webkit-text-size-adjust: 100%;
-    -moz-text-size-adjust: 100%;
-    -ms-text-size-adjust: 100%;
-}
-
-#home-button {
-    cursor: pointer;
-    margin: 5px;
-    text-align: center;
-}
-.center {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-}
-
-.upload {
-    border: 1px solid;
-    width: 100px;
-    height: 400px;
-}
-.image-container {
-    border: 1px solid;
-    width: 400px;
-    height: 400px;
-}
